Proyecto 1/requirements.txt
pandas==2.1.4
plotly_express==0.4.1
streamlit==1.29.0